fix: ci

+++ src/SimplevotingSystem.sol
@@ -5,7 +5,6 @@ import {Ownable} from "openzeppelin-contracts/contracts/access/Ownable.sol";
 import {AccessControl} from "openzeppelin-contracts/contracts/access/AccessControl.sol";
 
 contract SimpleVotingSystem is Ownable, AccessControl {
-    
     // --- ROLES ---
     bytes32 public constant FOUNDER_ROLE = keccak256("FOUNDER_ROLE");
     bytes32 public constant WITHDRAWER_ROLE = keccak256("WITHDRAWER_ROLE");
@@ -19,25 +18,25 @@ contract SimpleVotingSystem is Ownable, AccessControl {
     }
 
     struct Candidate {
-        uint id;
+        uint256 id;
         string name;
-        uint voteCount;
-        uint fundsReceived;
+        uint256 voteCount;
+        uint256 fundsReceived;
     }
 
     WorkflowStatus public workflowStatus;
-    uint public voteStartTime;
+    uint256 public voteStartTime;
 
-    mapping(uint => Candidate) public candidates;
+    mapping(uint256 => Candidate) public candidates;
     mapping(address => bool) public voters;
-    uint[] private candidateIds;
+    uint256[] private candidateIds;
 
     // --- EVENTS ---
     event WorkflowStatusChange(WorkflowStatus previousStatus, WorkflowStatus newStatus);
-    event CandidateRegistered(uint id, string name);
-    event Voted(address voter, uint candidateId);
-    event FundReceived(address founder, uint candidateId, uint amount);
-    event FundsWithdrawn(address withdrawer, uint amount);
+    event CandidateRegistered(uint256 id, string name);
+    event Voted(address voter, uint256 candidateId);
+    event FundReceived(address founder, uint256 candidateId, uint256 amount);
+    event FundsWithdrawn(address withdrawer, uint256 amount);
 
     constructor() Ownable(msg.sender) {
         workflowStatus = WorkflowStatus.REGISTER_CANDIDATES;
@@ -50,7 +49,7 @@ contract SimpleVotingSystem is Ownable, AccessControl {
     function setWorkflowStatus(WorkflowStatus _newStatus) public onlyOwner {
         WorkflowStatus previousStatus = workflowStatus;
         workflowStatus = _newStatus;
-        
+
         if (_newStatus == WorkflowStatus.VOTE) {
             voteStartTime = block.timestamp;
         }
@@ -64,16 +63,16 @@ contract SimpleVotingSystem is Ownable, AccessControl {
         require(workflowStatus == WorkflowStatus.REGISTER_CANDIDATES, "Candidates registration is not open");
         require(bytes(_name).length > 0, "Candidate name cannot be empty");
 
-        uint candidateId = candidateIds.length + 1;
+        uint256 candidateId = candidateIds.length + 1;
         candidates[candidateId] = Candidate(candidateId, _name, 0, 0);
         candidateIds.push(candidateId);
-        
+
         emit CandidateRegistered(candidateId, _name);
     }
 
     // --- FONCTIONNALITÉS FINANCIÈRES (FOUNDER & WITHDRAWER) ---
 
-    function fundCandidate(uint _candidateId) public payable onlyRole(FOUNDER_ROLE) {
+    function fundCandidate(uint256 _candidateId) public payable onlyRole(FOUNDER_ROLE) {
         require(workflowStatus != WorkflowStatus.COMPLETED, "Workflow is completed");
         require(_candidateId > 0 && _candidateId <= candidateIds.length, "Invalid candidate ID");
         require(msg.value > 0, "Amount must be greater than 0");
@@ -84,10 +83,10 @@ contract SimpleVotingSystem is Ownable, AccessControl {
 
     function withdraw() public onlyRole(WITHDRAWER_ROLE) {
         require(workflowStatus == WorkflowStatus.COMPLETED, "Workflow is not completed yet");
-        uint balance = address(this).balance;
+        uint256 balance = address(this).balance;
         require(balance > 0, "No funds to withdraw");
 
-        (bool sent, ) = payable(msg.sender).call{value: balance}("");
+        (bool sent,) = payable(msg.sender).call{value: balance}("");
         require(sent, "Failed to send Ether");
 
         emit FundsWithdrawn(msg.sender, balance);
@@ -95,7 +94,7 @@ contract SimpleVotingSystem is Ownable, AccessControl {
 
     // --- VOTE ---
 
-    function vote(uint _candidateId) public {
+    function vote(uint256 _candidateId) public {
         require(workflowStatus == WorkflowStatus.VOTE, "Voting session is not open");
         require(block.timestamp >= voteStartTime + 1 hours, "Voting starts 1 hour after session opening");
         require(!voters[msg.sender], "You have already voted");
@@ -103,22 +102,22 @@ contract SimpleVotingSystem is Ownable, AccessControl {
 
         voters[msg.sender] = true;
         candidates[_candidateId].voteCount += 1;
-        
+
         emit Voted(msg.sender, _candidateId);
     }
 
     // --- VIEW FUNCTIONS ---
 
-    function getTotalVotes(uint _candidateId) public view returns (uint) {
+    function getTotalVotes(uint256 _candidateId) public view returns (uint256) {
         require(_candidateId > 0 && _candidateId <= candidateIds.length, "Invalid candidate ID");
         return candidates[_candidateId].voteCount;
     }
 
-    function getCandidatesCount() public view returns (uint) {
+    function getCandidatesCount() public view returns (uint256) {
         return candidateIds.length;
     }
 
-    function getCandidate(uint _candidateId) public view returns (Candidate memory) {
+    function getCandidate(uint256 _candidateId) public view returns (Candidate memory) {
         require(_candidateId > 0 && _candidateId <= candidateIds.length, "Invalid candidate ID");
         return candidates[_candidateId];
     }
